add libreboot static library
this is a simple library that implements the "reboot_wrapper()"
call, which allows some nifty hax to happen before reboot for
devices that require it, without the need for identical code to
be maintained in multiple locations.

the idea is that ALL PARTS OF ANDROID will use this library to
reboot the system.

Change-Id: Id2c17191ebeb3bb171463f5e6683d6fdf2fff795